Move custom LDAPBackend to contrib.auth
